Refactor and cleanup
Entity and SceneNode moved to RenderComponent
