+++ P3/ProyectoBase/Src/Engine/3DObjectRC.cpp
@@ -1,12 +1,11 @@
 #include <json.h>
 
-#include "TridimensionalObject.h"
+#include "3DObjectRC.h"
 #include "Entity.h"
 #include "Factory.h"
 #include "FactoriesFactory.h"
 #include "ComponentsManager.h"
 #include "TransformComponent.h"
-#include "OgreVector3.h"
 #include "OgreEntity.h"
 #include "OgreSDLContext.h"
 #include "OgreSceneManager.h"
@@ -14,23 +13,13 @@
 // COMPONENT CODE
 TridimensionalObject::TridimensionalObject() {}
 
-TridimensionalObject::~TridimensionalObject() {
-	_entity = nullptr;
-	_sceneNode = nullptr;
-}
+TridimensionalObject::~TridimensionalObject() {}
 
 void TridimensionalObject::render() {}
 
 void TridimensionalObject::setMaterial(std::string material) {
 	_entity->setMaterialName(material);
 }
-
-// GETTERS AND SETTERS
-Ogre::Entity* TridimensionalObject::getOgreEntity() { return _entity; }
-void TridimensionalObject::setOgreEntity(Ogre::Entity* e) {	_entity = e; }
-
-Ogre::SceneNode* TridimensionalObject::getSceneNode() {	return _sceneNode; }
-void TridimensionalObject::setSceneNode(Ogre::SceneNode* sn) { _sceneNode = sn; }
 
 // FACTORY INFRASTRUCTURE
 class TridimensionalObjectFactory : public ComponentFactory {
@@ -49,11 +38,13 @@ public:
 		tridimensionalObject->getSceneNode()->attachObject(tridimensionalObject->getOgreEntity());
 
 		TransformComponent* transform = dynamic_cast<TransformComponent*>(father->getComponent("TransformComponent"));
-		Ogre::Vector3 ori = transform->getOrientation();
 		tridimensionalObject->getSceneNode()->setPosition(transform->getPosition());
-		//tridimensionalObject->getSceneNode()->setOrientation(0, ori.x, ori.y, ori.z);
 		tridimensionalObject->getSceneNode()->setScale(transform->getScale());
-		
+
+		// Pendiente de unificar un metodo con las fisicas y otros que necesiten rotacion
+		//Ogre::Vector3 ori = transform->getOrientation();
+		//tridimensionalObject->getSceneNode()->setOrientation(0, ori.x, ori.y, ori.z);
+
 		componentManager->addRC(tridimensionalObject);
 		return tridimensionalObject;
 	};
