+++ P3/ProyectoBase/Src/Engine/3DObjectRC.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <string>
+#include "RenderComponent.h"
+
+class TridimensionalObject : public RenderComponent {
+
+public:
+	TridimensionalObject();
+	virtual ~TridimensionalObject();
+
+	virtual void render();
+	void setMaterial(std::string material);
+};
+
